HV: refine vmptable for partition mode
Merge separated mptable.c of BOARDs to unified vmptable.c;

Tracked-On: #2291

Acked-by: Eddie Dong <eddie.dong@intel.com>
