+++ hypervisor/partition/apl-mrb/vm_description.c
@@ -168,7 +168,6 @@ struct vm_description_array vm_desc_partition = {
 						console=ttyS2 no_timer_check ignore_loglevel log_buf_len=16M \
 						consoleblank=0 tsc=reliable xapic_phys",
 				.vpci_vdev_array = &vpci_vdev_array1,
-				.mptable = &mptable_vm1,
 			},
 
 			{
@@ -183,7 +182,6 @@ struct vm_description_array vm_desc_partition = {
 						console=ttyS2 no_timer_check ignore_loglevel log_buf_len=16M \
 						consoleblank=0 tsc=reliable xapic_phys",
 				.vpci_vdev_array = &vpci_vdev_array2,
-				.mptable = &mptable_vm2,
 				.lapic_pt = true,
 			},
 		}
