+++ hypervisor/partition/cb2_dnv/vm_description.c
@@ -202,7 +202,6 @@ struct vm_description_array vm_desc_partition = {
 						"console=ttyS0 no_timer_check ignore_loglevel log_buf_len=16M "\
 						"consoleblank=0 tsc=reliable xapic_phys  apic_debug",
 				.vpci_vdev_array = &vpci_vdev_array1,
-				.mptable = &mptable_vm1,
 				.lapic_pt = true,
 			},
 
@@ -218,7 +217,6 @@ struct vm_description_array vm_desc_partition = {
 						"console=ttyS0 no_timer_check ignore_loglevel log_buf_len=16M "\
 						"consoleblank=0 tsc=reliable xapic_phys apic_debug",
 				.vpci_vdev_array = &vpci_vdev_array2,
-				.mptable = &mptable_vm2,
 				.lapic_pt = true,
 			},
 		}
